fix edit model button
